requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.2
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
